+++ packages/integration/src/index.ts
@@ -8,6 +8,8 @@ export { getPackageInfo } from './packageInfo';
 export { compile, vanillaExtractFilescopePlugin } from './compile';
 export { hash } from './hash';
 export { addFileScope } from './addFileScope';
+export { serializeCss, deserializeCss } from './serialize';
+
 export * from './filters';
 
 export type { IdentifierOption } from './processVanillaFile';