+++ test-helpers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vanilla-extract-private/test-helpers
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fixtures/sprinkles@0.0.6
+
 ## 0.0.7
 
 ### Patch Changes