+++ tests/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vanilla-extract-private/tests
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`81a3d98`](https://github.com/seek-oss/vanilla-extract/commit/81a3d98b3fb36660406a59049e1cd464418b00fc)]:
+  - @vanilla-extract/sprinkles@1.4.0
+  - @vanilla-extract-private/test-helpers@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes